Effective Touch Action when <body> is the viewport defining element
When <body> is deemed to be the viewport defining element <body>'s overflow needs to be set to visible. Because of this, during effective touch action calculation we consider <body> to be a scroller when it is not.

The fix attempts to detect when <body> is deemed the viewport defining element and correct the effective touch action calculation.

Bug: 1031745
Change-Id: Ie9a3c966af78fdbe0dc3e57f3f192aca9110fd0e